% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminated_rand.R
\name{elimination_rand}
\alias{elimination_rand}
\title{Title Determine the players that will be eliminated}
\usage{
elimination_rand(num_players)
}
\arguments{
\item{num_players}{Number of players participating in the game}
}
\value{
list of players eliminated
}
\description{
Title Determine the players that will be eliminated
}
\examples{
elimination_rand(num_players = 12)
}
