% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_NE.R
\name{game_NE}
\alias{game_NE}
\title{Nash Equilibrium}
\usage{
game_NE(num_of_firms, K = "none")
}
\arguments{
\item{num_of_firms}{Number of firms in the game.}

\item{K}{Amount available of the Commons. Default is number of firms + 1.}
}
\value{
A written statement explaining the strategies that lead to Nash Equilibrium.
}
\description{
Nash Equilibrium
}
\examples{
game_NE(2)
}
